pyproject.toml
[project]
name = "flourishing-nest-ad-hoc-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "pdfplumber>=0.11.4",
    "psycopg[binary]>=3.2",
    "python-dotenv>=1.2.1",
]
